Split multi-scope queries into separate queries
By using `distinct.pluck(:id)`, our Ruby server gets a list of IDs for the requested object. This ensures the scope listed before is executed first and the result (the resulting IDs) is then passed to the next query. Similar to the previous commit, this change ensures we are not repeatedly performing the same subquery over and over again (and thus improves query performance).

Remember: A where clause is evaluated per row, and so is a subquery.