Release runner in case of errors
Previously, a runner was not released in case of an `Runner::Error`. Especially when a runner was not found (i.e., when executing an arbitrary command), a `Runner::Error::RunnerNotFound` exception was thrown. Then, the automatic request for a new runner failed, since the previous runner was still in use.