Use current path for form_filter search partial
Previously, the search always used the index page of the model (through `@search.klass`). Usually, this works fine, but for the RfC action, we want to have a search for the pre-filtered actions, too. Since we are using the `request.path` already to clear the filter and didn't notice any issues with that, we also use this approach for the actual search. The backend is ready to accept searches with the two additional RfC endpoints.