Only allow stopping the current code execution when started
When a code execution is started, we shortly show a spinner on the Run button before it changes to Stop, allowing learners to abort the current code execution. However, previously, the Stop button was shown too early namely when the WebSocket connection was getting initialized. When it was immediately pressed after becoming available, it didn't had any effect.

With the changes of this commit, we delay showing the Stop button, so that it will be shown only once the runner is actually beginning with the code execution (as indicated by the `container_running` status).