PyUnitAdapter: Handle an edge case with aborted tests
We recently added an exercise to CodeOcean that skipped code execution during the `setUpClass` phase. This resulted in the following output:

```
Ran 0 tests in 0.001s

FAILED (errors=2)
```

As a result, we calculated `count=0`, `failed=2` (and through TestingFrameworkAdapter#augment_output) `passed=-2`. Obviously, this doesn't make sense.

As the output is generated directly through PyUnit, we cannot change it and hence implement a workaround in the adapter to "fake" the number of total tests executed. Not super nice, but working.