Introduce a new DeadlineScope for Submissions
With this scope, we extend the submissions shown to teachers. Now, a teacher has always access to the latest scored submission and the submission with the highest score (in addition to all "submit" submissions). Furthermore, for these two categories (latest / highest scored), we also respect the three deadlines (before submission deadline, within grace period, after late submission deadline) and show a respective submission.

The SubmissionsController#index has not been extended with this new scope, since the performance is too bad (and it is not used frequently).