Disable ACE worker for RfC comment area
Otherwise, annotations would be overwritten by an ACE worker, if available. This bug was only present for a few file types (incl. JavaScript, JSON, PHP, CoffeeScript, CSS, XQuery) as listed in the official docs [1]. If this bug was triggered and no syntax error was present in the learner's file, this would simply "remove" the comment annotations and prevent accessing existing comments written for the file.

Closes #2751

[1] https://github.com/ajaxorg/ace/wiki/Syntax-validation