Remove rails-erb-loader and unify JS-generated files
We have two integrations between JavaScript and Rails:
* js-routes
* i18n-js

Both allow using the same interface in JavaScript as their Ruby pendant. Previously, the setup of both was different and somewhat "hidden" (with temporary files). Now, we make the setup more obvious and put the generated files in a subdirectory of the main `app/javascript` folder.

For the Raketask integration, we use the respective Ruby methods directly, rather than invoking a subprocess (improves performance).