Enforce uniqueness constraint on external_id of external_users
Previously, we lacked such a constraint (cf. #2705), causing duplicated external users in production. Probably, they were caused by a double-click on the "Launch exercise", starting a race condition in two tabs to launch the exercise. I assume that both requests succeeded and each created a new external user.

Obviously, an external ID should be unique per consumer, so that this error must not happen. We fix it by adding the respective database constraint, and a model validation in Rails (preventing similar issues in the future). For the existing issues, we provide a script, allowing to merge two external user accounts into a single user.