Enforce uniqueness constraint on external_id of study_groups
Similar to e9ab9f84, an external_id of study_groups should be unique. In addition to the previously-configured PostgreSQL constraint, we also add a Rails model validation.

Furthermore, we adjust the PostgreSQL index by specifying `nulls_not_distinct: true`. This is beneficial, since the external_id of study_groups CAN be NULL (in exactly one case, namely for the default study group per consumer). In all other cases, the external_id MUST be set (which this index enforces).

Finally, we use this opportunity to update the migrate_study_group.sql script in case it is needed.