Handle additional tables in migrate_exercise.sql
Since adding the script in 2020 with 7ab4c63b, many tables have changed. Hence, we use this opportunity to update the script, allowing it to be used with the current database design. Furthermore, we also keep more data if possible, smoothing a migration.

Unfortunately, migrating exercises got a bit more complicated due to programming groups, community solutions, and tips.

Since we didn't require a migration for these yet, we just added a safeguard for now (acknowledging that some foreign key constraints might still be missing).