Specs: Use default study group for internal users
When creating a teacher (or admin) through the respective factory, they were not yet automatically assigned to the respective default study group. This commit changes said behavior.

The actual creation process within the UI remains unchanged, and it is still required to choose the appropriate study groups manually.

Part of #2986