Specs: Wait for JavaScript actions on /implement
The Score system specs are very tricky and error-prone with the current setup. Mostly, this is due to the headless Chrome browser used on CI, which has a weird race condition preventing a regular use. Until said issue is resolved, we add additional expectations (to wait for the page to finish loading).

See https://github.com/SeleniumHQ/selenium/issues/15273