JavaScript: Show editor tooltips only on hover
This change might be needed to prevent issues with open tooltips when navigating a page. After migrating to Turbo, we see a few occasions in Bootstrap accessing an active tooltip when navigating to another page (i.e., when finalizing an exercise).

Depending on whether the issue is resolved, we might need to investigate further.

Aims to fix CODEOCEAN-FRONTEND-D9