Prevent page unloading in case of unknown editor
Recently, we observed an issue with a missing `editor_for_file`, which caused a `turbo:visit` event and unloaded the editor. By preventing the bug, the page is not unloaded and continuous to work as expected (despite clicking on the line won't do anything, which is less severe).

Fixes CODEOCEAN-FRONTEND-DR