fix: Avoid reports of legitimate content
Social media platforms use `Melden` to report inappropriate content.
In the teaching context, it is common in German to report yourself
if you have a question (sich melden). It appears that students report
content to get attention from the teacher because no reported content
includes inappropriate content.

It is the hope that `Inhalt melden` makes it clearer that this is a
content moderation function.

The popup message also asks the student if he wants to report the
content. This should make it clearer that it is not "sich melden".
Due to the Turbo.js migration, this message is not always displayed,
or the students ignore the message. #3021 might also cause\contribute
to the issue.