fix: Faster theme loading to avoid flickering (#3062)
The `tubo:load` event is faster than the migration event. The color changing requires no sprocket code and can be controlled without the migration event. This removes the flickering in dark mode when navigating the page.

This change is possible because the theme-changing logic was moved to Webpacker and loads before Sprockets. Please see #3025

Related to #3040